Refactor: 크롤러 유틸리티 병합
