+++ src/crawling/crawler.service.ts
@@ -1,9 +1,21 @@
 import { Injectable, Logger } from "@nestjs/common";
 import * as puppeteer from 'puppeteer';
+import PQueue from '@esm2cjs/p-queue';
 
 @Injectable()
 export class CrawlerService {
   private readonly logger = new Logger(CrawlerService.name);
+
+  private readonly queue = new PQueue({ concurrency: 5 });
+  // 크롤링 큐 설정
+  async setQueue<T>(tasks: Array<() => Promise<T>>): Promise<T[]> {
+    return Promise.all(tasks.map((task) => this.queue.add(task)));
+  }
+  // 지연 시간 설정
+  async setDelay(min: number = 1000, max: number = 3000) {
+    const randomTime = Math.floor(Math.random() * (max - min + 1)) + min;
+    return new Promise(resolve => setTimeout(resolve, randomTime));
+  }
   // 크롤링 모듈 선언 및 초기화
   private browser: puppeteer.Browser;
   // 브라우저 초기화
